refactor(worker): seperate loading and prediction queues
